Create departments

+++ database/migrations/2016_03_14_212646_create_order_details_table.php
@@ -19,7 +19,7 @@ class CreateOrderDetailsTable extends Migration
              $table->string('product_desc');
              $table->tinyInteger('line');
              $table->smallInteger('qty');
-             $table->integer('price');
+             $table->decimal('price', 12, 2);
              $table->timestamps();
 
              $table->index('product_code');
